+++ packages/router/src/routes/api/icon_doc/all_bin_md5.rs
@@ -1,0 +1,17 @@
+use anyhow::Result;
+
+use axum::{extract::Json, http::StatusCode, response::IntoResponse};
+
+use crate::middlewares::ExtractAuthInfo;
+
+/// 获取所有图标信息的 MD5
+/// GET /icon_doc/all_bin_md5
+#[tracing::instrument(skip(auth))]
+pub async fn all_bin_md5(
+    ExtractAuthInfo(auth): ExtractAuthInfo,
+) -> Result<impl IntoResponse, (StatusCode, String)> {
+    match _functions::functions::api::icon_doc::do_all_bin_md5(auth, serde_json::json!({})).await {
+        Ok(v) => Ok((StatusCode::OK, Json(v))),
+        Err(e) => Err((StatusCode::INTERNAL_SERVER_ERROR, format!("{}", e))),
+    }
+}